Convert to DGS

+++ src/main/java/jimnorth1982/javagqlshowcase/model/RentalAgreementInput.java
@@ -3,5 +3,5 @@ package jimnorth1982.javagqlshowcase.model;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record RentalAgreementInput(Long toolId, BigDecimal discount, LocalDate checkOutDate, int rentalDays) {
+public record RentalAgreementInput(Integer toolId, BigDecimal discount, LocalDate checkOutDate, Integer rentalDays) {
 }
