+++ src/main/java/jimnorth1982/javagqlshowcase/repo/RentalAgreementRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.repository.Repository;
 
 import java.util.List;
 
-public interface RentalAgreementRepository extends Repository<RentalAgreement, Long> {
-    public RentalAgreement findById(Long id);
+public interface RentalAgreementRepository extends Repository<RentalAgreement, Integer> {
+    public RentalAgreement findById(Integer id);
 
     public List<RentalAgreement> findAll();
